pdf-bot/pyproject.toml
[tool.poetry]
name = "pdf-bot"
version = "0.1.0"
description = ""
authors = ["HypedAugust <44738421+HypedAugust@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langchain = "^0.2.11"
langchain-openai = "^0.1.19"
chromadb = "^0.5.5"
pypdf = "^4.3.1"
gradio = "^4.39.0"
gradio-pdf = "^0.0.11"
langchain-community = "^0.2.10"
langchain-text-splitters = "^0.2.2"
langchain-chroma = "^0.1.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
